test(apc): fold #1562 join coverage into existing test modules
Move staggered join / last-layer warm policy tests out of a new file
into test_apc_quantized.py and test_batch_quantized_cache.py, matching
review guidance on #1568 (prefer existing test scripts).
